LineFollowingwithPololu_QTR8_ArraySensor
Work in Progress: Sensor will need to adjust and correct the driving of the robot in the real time and if the line is not detected the robot should search for the line itself and once it detects follow it when driving using the data from sensor. Reading and Calibration/Re-calibration part when using the sensor is nearly complete, but it still needs testing. That's the task for the WorldSkills 2019 Competition in Robotics.